Refactor: Change Customer Entity Name to Member
