--- src/main/java/personal/hdproject/customer/dao/RefreshTokenRepository.java
@@ -1,10 +1,0 @@
-package personal.hdproject.customer.dao;
-
-public interface RefreshTokenRepository {
-
-	void save(Long id, String refreshToken);
-
-	void delete(String id);
-
-	String getRefreshToken(Long id);
-}
