--- src/main/java/personal/hdproject/customer/dao/common/CustomerCustomRepository.java
@@ -1,8 +1,0 @@
-package personal.hdproject.customer.dao.common;
-
-public interface CustomerCustomRepository {
-
-	Long updateNickname(Long customerId, String nickname);
-
-	Long updatePhone(Long customerId, String phone);
-}
