--- src/main/java/personal/hdproject/customer/domain/Grade.java
@@ -1,5 +1,0 @@
-package personal.hdproject.customer.domain;
-
-public enum Grade {
-	BASIC, VIP
-}
