--- src/main/java/personal/hdproject/customer/service/request/SignInCustomerServiceRequest.java
@@ -1,17 +1,0 @@
-package personal.hdproject.customer.service.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class SignInCustomerServiceRequest {
-
-	private final String email;
-	private final String encryptedPassword;
-
-	@Builder
-	private SignInCustomerServiceRequest(String email, String encryptedPassword) {
-		this.email = email;
-		this.encryptedPassword = encryptedPassword;
-	}
-}
