+++ src/main/java/personal/hdproject/member/dao/RefreshTokenRepository.java
@@ -1,0 +1,10 @@
+package personal.hdproject.member.dao;
+
+public interface RefreshTokenRepository {
+
+	void save(Long id, String refreshToken);
+
+	void delete(String id);
+
+	String getRefreshToken(Long id);
+}
