+++ src/main/java/personal/hdproject/member/dao/common/MemberCustomRepository.java
@@ -1,0 +1,8 @@
+package personal.hdproject.member.dao.common;
+
+public interface MemberCustomRepository {
+
+	Long updateNickname(Long memberId, String nickname);
+
+	Long updatePhone(Long memberId, String phone);
+}
