+++ src/main/java/personal/hdproject/member/domain/Grade.java
@@ -1,0 +1,5 @@
+package personal.hdproject.member.domain;
+
+public enum Grade {
+	BASIC, VIP
+}
