+++ src/main/java/personal/hdproject/util/interceptor/WebConfig.java
@@ -1,23 +1,25 @@
 package personal.hdproject.util.interceptor;
 
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import lombok.RequiredArgsConstructor;
-import personal.hdproject.util.interceptor.handler.CustomerLoginCheckHandler;
+import personal.hdproject.util.interceptor.handler.MemberLoginCheckHandler;
 
+@Profile("product")
 @Component
 @RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
-	private final CustomerLoginCheckHandler customerLoginCheckHandler;
+	private final MemberLoginCheckHandler memberLoginCheckHandler;
 
 	@Override
 	public void addInterceptors(InterceptorRegistry registry) {
-		registry.addInterceptor(customerLoginCheckHandler)
-			.excludePathPatterns("/api/v1/customer/sign-in")
-			.excludePathPatterns("/api/v1/customer")
+		registry.addInterceptor(memberLoginCheckHandler)
+			.excludePathPatterns("/api/v1/member/sign-in")
+			.excludePathPatterns("/api/v1/member")
 			.addPathPatterns("/**");
 	}
 }
