+++ src/main/resources/application.yml
@@ -1,9 +1,11 @@
 spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/hot_dealicious
+    url: jdbc:mysql://localhost:3306/hot_deal
     username: root
     password: 1234
+
+---
 
 redis:
   refresh_token:
