+++ src/test/resources/application.yml
@@ -1,4 +1,8 @@
 spring:
+  config:
+    activate:
+      on-profile: test
+
   datasource:
     driver-class-name: org.h2.Driver
     url: jdbc:h2:mem:test
@@ -10,3 +14,16 @@ spring:
 
 logging.level:
   org.hibernate.sql: debug
+
+---
+
+redis:
+  refresh_token:
+    host: localhost
+    port: 6379
+
+jwt:
+  header: x-auth-token
+  secret_key: 7c8d3c13bd32be41c39655f87737422920037a5ccf569bf8484511cf5241ceec
+  at-expired_minutes: 5
+  rt-expired_days: 90
